requirements.txt
requests
discord
spotipy
python-dotenv
termcolor
numpy
pillow
pyyaml
opencv-python
moviepy
